changed to unsigned longs; addressed all warnings

+++ edge_to_graph.c
@@ -12,8 +12,8 @@
 // List_of_offset_deg_pairs [2N]
 // N lists of D vertices
 
-int N,M,D;
-unsigned int* map;
+unsigned long N,M,D;
+unsigned long* map;
 
 /*void add_edge(int u, int v){
 	//adds the edges to u and v's lists
@@ -55,9 +55,9 @@ int main(int argc, char *argv[]) {
 	//int PAGE_SIZE = 4096;
 	
 	//TODO Replaced this to get the page size
-	long PAGE_SIZE = sysconf(_SC_PAGESIZE);
+	unsigned long PAGE_SIZE = sysconf(_SC_PAGESIZE);
 
-	int header_size = sizeof(unsigned int)*(3 + 2*N); //N, M, D(3) [o,d]*N in 
+	unsigned long header_size = sizeof(unsigned long)*(3 + 2*N); //N, M, D(3) [o,d]*N in 
 	printf("The header is (%d x 2 + 5) X 4 = %d bytes\n", N, header_size);
 	printf("The pages are %d bytes\n", PAGE_SIZE);
 	
@@ -66,38 +66,38 @@ int main(int argc, char *argv[]) {
 	double hpages = (double)header_size/PAGE_SIZE;
 	printf("There are %f float header pages\n", hpages);
 
-	int num_hpages = ceil(hpages);
+	unsigned long num_hpages = ceil(hpages);
 	printf("Which is %d integer header pages\n", num_hpages);
 	
-	int padding = (num_hpages * PAGE_SIZE) - header_size; //in bytes
+	unsigned long padding = (num_hpages * PAGE_SIZE) - header_size; //in bytes
 	
 	
-	int D_up;
+	unsigned long D_up;
 	double junk = pow(2,ceil(log2(D)));
 	D_up = junk;
 
 	printf("The upper bound of D is %d\n", D_up);
 	
-	int node_data_size = sizeof(unsigned int)*N*D_up; //in bytes
+	unsigned long node_data_size = sizeof(unsigned long)*N*D_up; //in bytes
 
 	
 	//stretch the file in bytes
-	size_t length = num_hpages*PAGE_SIZE + sizeof(unsigned int)*D_up + node_data_size; // N, M, D; [n,o]*N + padding; burn row [N][D]
+	size_t length = num_hpages*PAGE_SIZE + sizeof(unsigned long)*D_up + node_data_size; // N, M, D; [n,o]*N + padding; burn row [N][D]
 	if (lseek(fd, length+1, SEEK_SET) == -1){
 		close(fd);
 		perror("Error on lseek call to stretch the file");
 		exit(EXIT_FAILURE);
 	} //position at end of file
 	
-	double dpages = ((sizeof(unsigned int)*D_up + node_data_size))/(double)PAGE_SIZE;
+	double dpages = ((sizeof(unsigned long)*D_up + node_data_size))/(double)PAGE_SIZE;
 	printf("There are %f double node pages\n", dpages);
 	
-	int num_dpages = ceil(dpages);
+	unsigned long num_dpages = ceil(dpages);
 	printf("Which is %d int node pages\n", num_dpages);
 	
-	int num_tpages = num_hpages + num_dpages;
+	unsigned long num_tpages = num_hpages + num_dpages;
 
-	printf("The length of this file is based on header %d + padding %d + node data %d + burned row %d = %d\n", header_size, padding, node_data_size, sizeof(int)*D_up, length);
+	printf("The length of this file is based on header %d + padding %d + node data %d + burned row %d = %d\n", header_size, padding, node_data_size, sizeof(unsigned long)*D_up, length);
 	
 	printf("The length of the file is %d, requiring %d pages\n", length, num_tpages);
 
@@ -124,9 +124,9 @@ int main(int argc, char *argv[]) {
 
 	rewind(fp);
 	
-	int buf[3]={(unsigned int)N, (unsigned int)0, (unsigned int)D_up};
+	unsigned long buf[3]={(unsigned long)N, (unsigned long)0, (unsigned long)D_up};
 	
-	fwrite(buf, sizeof(unsigned int), 3, fp);
+	fwrite(buf, sizeof(unsigned long), 3, fp);
 	
 	/*
 	if (fprintf(fp,"%d",(int)0) < 0 ) {
@@ -169,14 +169,14 @@ int main(int argc, char *argv[]) {
 	}
 	
 	//initialize nodes with offset and degree 0
-	for(unsigned int i=0; i<my_graph->N ;i++){
+	for(unsigned long i=0; i<my_graph->N ;i++){
 		my_graph->g[3+2*i]=i;
 		my_graph->g[3+2*i+1]=0;
 	}
 
 	print_graph(my_graph);
 	
-	int u,v;
+	unsigned long u,v;
 	while(fscanf(fe, "%d %d", &u, &v)==2){
 		printf("Adding edge %d %d in e2g\n", u, v);
 		add_edge(my_graph, u, v);
